pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blockrun-llm-xrpl"
version = "0.1.0"
description = "BlockRun SDK for XRPL - Pay-per-request AI via x402 with RLUSD"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "BlockRun", email = "hello@blockrun.ai" }
]
keywords = ["llm", "ai", "x402", "xrpl", "rlusd", "micropayments", "openai", "claude", "gemini"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "httpx>=0.25.0",
    "xrpl-py>=4.0.0",
    "pydantic>=2.0.0",
    "x402-xrpl>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=24.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://blockrun.ai"
Documentation = "https://docs.blockrun.ai"
Repository = "https://github.com/BlockRunAI/blockrun-llm-xrpl"

[tool.hatch.build.targets.wheel]
packages = ["blockrun_llm_xrpl"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
strict = true
